terms: add pawn phalanx
Pretty simple term: apply a bonus if pawns are adjecent to each other.
This means that the pawns control N+1 squares in front of them and that
they can potentially support each when being pushed.

NOTE: this works as we also have safety / attacks terms already, so we
don't just blindly put the pawns next to each other.

Bench 2955703

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>