searcher: add `isChecked` to negamax eval
Our static evaluation is quite off when we're checked. It's a "lost"
position no matter what, so we should reflect that in the evaluation.

The same goes for the "isImproving" - we're not in a better position if
we're checked.

This fixes missing mating lines and in general a safer search.

NOTE: we already do the same in qsearch.

Bench 3113053

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>