see_swap: correct en-pessant
Prior to this commit we simply ignored the fact that en-pessant also
could be evaluated as SEE.

In fact we actually always assumed that an en-pessant would always be a
good noisy move. Instead we can realtively easy add SEE handling to
en-pessant captures and actually reflect them correctly in our move
ordering.

Bench 2904972

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>