searcher: add extended futility pruning
This commit adds initial static forward pruning.

The first pruning technique is called "extended futility pruning" (EFP).
The theory here is that after searching a certain amount of nodes, if
the evaluation is marginally worse than alpha then we can assume that
we're not going to improve that with "bad moves".

The "bad moves" are currently filtering out quiets and bad promotions.
We could be more aggresive and also filter bad captures, but let's keep
it a bit safer for now.

This scheme also makes it easy to apply more types of static forward
pruning.

Bench 2820965

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>