move_picker: swap-and-reduce for TT move
Instead of leaving null 'holes' in the move picker, we move picked
moves to the back of the array of all moves and reduce the size
accordingly.

Bench has changed, because quiet moves at the root have no history, and
are therefore picked in their order of generation. So changing the move
order has large effects down the search.

Bench 1446371