searcher: add see pruning
SEE pruning is an attempt at pruning branches by evaluating the
worst/best case scenario in a simplified way.

If the SEE score is significantly low then we can assume that the given
branch is not worth searching.

This implementation is very similar to our capture score but here we
just compute a balance - and quiet moves are also accepted.

Bench 1209330