mask_tables: move "position tables" into mask tables file
This will make it possible to add more generic masks to the file as
well.
Also it was stored in evaluation context which is not correct.
Now it's stored as a core feature which should be able to be used by any
part of the engine.

This commit also cleans up the mask tables a bit.
Now there's less redundant code and we store all the mask tables in one
location.

Bench 1354847

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>