see_swap: remove custom pawn attack mask table
Before we had an odd mask table; a kinda reversed pawn attack mask where
we could look up the pawns based on where it would attack from for a
given position.

This is quite redundant as we already have a way to compute this.. ie.
just invert the attack (use white instead of black eg).

Bench 1415419

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>